Fix for Invalid URL Error in BaseURL
On second and subsequent calls to GetBaseURL the value for ADSLESetup."Account Name" was NULL due to ADSLE Setup not being initialised to the setup record. GetSingleton() as been moved out of the if clause to fix this. This was causing loads to hang at "In Process" on submission of jobs.